# [onert] Arithmetic kernels without fused activation