Introduced kernels for non-fused arithmetic operations: Add, Mul, Sub

ONE-DCO-1.0-Signed-off-by: Alexander Efimov <a.efimov@samsung.com>