@glistening 2020-10-04
@binarman I think you have the experimental number for our target model. Could you please write down how much improvement it bring? Please write `before` and `after`'s number of `nnpackage_run --nnpackage our_model -w 300 -r 1200` on your android aarch64 device.