@binarman 2020-10-06
@glistening 
I've tested changes again like this (`--num_runs=10000 --warmup_runs=10`):
```
for i in $(seq 20); do taskset -a 128 ./nnpackage_run --num_runs=10000 --warmup_runs=10 --nnpackage package 2>&1 | grep " MEAN"; sleep 3; done
```

Using large number of runs I can get more or less stable results.
In short I got 1.2096 ms -> 1.2011 ms (0.65% improvement :man_shrugging:)

Full times output are:
With patch: [1.205, 1.199, 1.197, 1.206, 1.203, 1.236, 1.206, 1.186, 1.197, 1.234, 1.169, 1.202, 1.207, 1.193, 1.190, 1.197, 1.198, 1.192, 1.199, 1.206]
Without patch: [1.221, 1.194, 1.192, 1.225, 1.230, 1.207, 1.201, 1.223, 1.203, 1.209, 1.191, 1.188, 1.223, 1.204, 1.209, 1.212, 1.195, 1.218, 1.236, 1.211]

P.s. I also found that some BIG cores are faster then others on my phone, I didn't get why this is happening yet, but at least they are stable. So do not be scared by ~1.2 ms vs ~0.4 ms on your device.