@binarman 2020-10-06
As @krayzemli wrote he included these changes in his PR, so we can close this one.